Settle on CS4 as the scripting models are notably different
I was trying to be able to use both CS4 and CS5.5,
as CS4 on my home machine doesn’t work properly.

But it does work enough to be able to generate the
pages with a script, which is all that matters (as
the pages can be opened and checked in CS5.5).